TST: update all `result_to_tuple=` callables to accept two arguments
Update all the `result_to_tuple=` callables to accept two arguments,
and remove the compatibility hack for one-argument callables, fixing
the test suite on PyPy3.11.  The hack did not account for
`inspect.signature(tuple)` including `*args, **keywords` on PyPy
which led to it being misdetected as not requiring the compatibility
wrapping.  Rather than add further complexity to the hack, let's just
update all the functions to accept two arguments.

In addition to running the test suite with the changes, I have also
tried replacing the original hack with an explicitly raised exception
to make sure that no `result_to_tuple=` are passed a single-parameter
function.

Fixes #23035.